feat(logs-server)!: remove query CLI verb, standardize on sqlite3 + logs_v
The query verb was already demoted in docs to "sanity-check only" — it can't
reach structured probe attributes (greps msg body) and costs 10x sqlite's
cold-boot. Doctrine is now: `sqlite3 -header -column .davstack/logs/<db>`
against logs_v with json_extract(attrs, '$.<key>'). Docs rewritten to lead
with this pattern. Bumps to 2.1.0 (staying in 2.x per project policy of
batching breaking changes early on).
