+++ packages/logs-server/README.md
@@ -28,25 +28,27 @@ pnpm exec logs-server serve
 logger.debug("user clicked save", { user_id: 42, run_id: "r-99" })
 ```
 
+Read the store directly with `sqlite3` against the `logs_v` view (flat `attrs` column strips the OTel `{value,type}` wrapper):
+
 ```bash
-$ logs-server query filter --grep "clicked save"
-2026-05-21T13:51:32  debug  app  r-99  user clicked save  {"user_id":42}
+sqlite3 -header -column .davstack/logs/default.db "
+  SELECT ts, msg, json_extract(attrs, '\$.user_id') AS user_id
+  FROM logs_v
+  WHERE json_extract(data, '\$.body') LIKE '%clicked save%'
+  ORDER BY ts;
+"
 ```
 
-For non-trivial cuts, query sqlite directly via the `logs_v` view (flat `attrs` column strips the OTel `{value,type}` wrapper):
-
-```sql
--- sqlite3 .davstack/logs/default.db
-SELECT ts, msg, json_extract(attrs, '$.user_id') AS user_id
-FROM logs_v
-WHERE json_extract(data, '$.body') LIKE '%clicked save%'
-ORDER BY ts;
+```
+ts          msg                user_id
+----------  -----------------  -------
+1716480923  user clicked save  42
 ```
 
 ## Docs
 
 - [docs/setup.md](./docs/setup.md) — config file, env vars, runtime selection
 - [docs/writing-logs.md](./docs/writing-logs.md) — transmitter setup per SDK
-- [docs/reading-logs.md](./docs/reading-logs.md) — sqlite schema, recipes, and CLI verb reference (sqlite is the recommended read path for any non-trivial diagnosis)
+- [docs/reading-logs.md](./docs/reading-logs.md) — sqlite schema, the `logs_v` view, and ready-to-paste recipes
 - [docs/transmitter-wiring.md](./docs/transmitter-wiring.md) — route a session's logs to its own DB via the `davstack-logs.db` attribute
 - [docs/session-views.md](./docs/session-views.md) — per-DB SQL views, the high-value follow-up to multi-DB routing
